requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-framebuf==1.6.6
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-ssd1306==2.12.18
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
cbor2==5.6.5
colorzero==2.0
gpiod==2.2.3
gpiozero==2.0.1
lgpio==0.2.2.0
luma.core==2.4.2
luma.oled==3.14.0
pillow==11.0.0
pyftdi==0.56.0
pyserial==3.5
pyusb==1.2.1
rpi-lgpio==0.6
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
setuptools==75.6.0
smbus==1.1.post2
smbus2==0.5.0
spidev==3.6
sysv_ipc==1.1.0
typing_extensions==4.12.2
